refactor(bridge): extract test-only presence emitters
