+++ whatsrust/lib/main.go
@@ -83,10 +83,6 @@ static void callMessageHandler(MessageHandler hdl, bool isSync, const Message* d
     hdl.callback(&copy, isSync, hdl.user_data);
 }
 
-static void callPresenceHandler(PresenceHandler hdl, JID from, bool unavailable, int64_t lastSeen) {
-	hdl.callback(from, unavailable, lastSeen, hdl.user_data);
-}
-
 typedef void (*HistorySyncCallback)(uint32_t, void*);
 typedef struct {
 	HistorySyncCallback callback;
@@ -562,24 +558,6 @@ func messageActionKindName(kind uint8) string {
 		return "edit"
 	}
 	return "delete"
-}
-
-//export C_TestEmitPresenceEvent
-func C_TestEmitPresenceEvent(from *C.char, unavailable C.bool, lastSeen C.int64_t) {
-	C.callPresenceHandler(presenceHandler, from, unavailable, lastSeen)
-}
-
-//export C_TestEmitPresenceEventsConcurrently
-func C_TestEmitPresenceEventsConcurrently(from *C.char, count C.uint32_t) {
-	var wait sync.WaitGroup
-	for index := uint32(0); index < uint32(count); index++ {
-		wait.Add(1)
-		go func(lastSeen uint32) {
-			defer wait.Done()
-			C.callPresenceHandler(presenceHandler, from, C.bool(false), C.int64_t(lastSeen))
-		}(index)
-	}
-	wait.Wait()
 }
 
 func main() {} // Required for CGO
